HalberdAttackLibrary class added
core functions updated
