+++ core/playbook/playbook_error.py
@@ -1,0 +1,3 @@
+class PlaybookError(Exception):
+    """Custom exception class for Playbook-related errors."""
+    pass
